Update go libraries (#97)
* Update go libraries

Updated kubernetes libraries to 1.28.6
Updated kubevirt go-client to 1.1.1
Updated containerized data importer to 1.58.1

Switched virtcli from using the one provided by kubevirt to
a generated one using gen-client. This allows us to be independent
of the kubernetes version used by kubevirt. Also generate go-client
for containerized data importer to be independent of their kubernetes
version as well.

Signed-off-by: Alexander Wels <awels@redhat.com>

* Fix generating clients to install gen-client
Fix linter errors

Signed-off-by: Alexander Wels <awels@redhat.com>

* Switch port-forward to using virtclt binary instead of
of having it as a library. No need to maintain the library
this way.

Moved port-forward to the before suite and after suite, so
it will only have to forward the ports once for the whole test
suite.

Added function that finds a free port to use for port-forwarding.

Signed-off-by: Alexander Wels <awels@redhat.com>

* Addressed review comments.

- Removed kubevirt go-client and vendoring.
- Removed unused function

Signed-off-by: Alexander Wels <awels@redhat.com>

---------

Signed-off-by: Alexander Wels <awels@redhat.com>