Fix default allowed not working (#109)
Fix bug where if no mapping exists but default is
set to allowed, it would reject snapshot requests
because it could not find the mapping in the
tenant volume snapshot class. This fix will return
a blank mapping which then leaves the volume
snapshot class for the infra empty relying on the
default being set in the infra cluster.

Signed-off-by: Alexander Wels <awels@redhat.com>