Ensure volume stays attached through reboots (#130)
* Ensure volume stays attached through reboots

While this is not stated explicitly in the CSI spec,
k8s expects ControllerPublished volumes to stay attached regardless of reboots.

Today, we lose the attachment if a reboot is quick enough that a
failover to a different node was not initiated, and there is no extra
invocation of the publish, resulting in an eternal failure.

We can align to this implicit guideline by hotplugging on the VM level:
https://kubevirt.io/user-guide/storage/hotplug_volumes/#retain-hotplugged-volumes-after-restart

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

* fix gitignoring the folder cmd/kubevirt-csi-driver

this is an issue with vscode since searching will
exclude that folder

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

* Adapt publish/unpublish for VM level hotplug

This is mainly achieved through changing the VM identification
mechanism to namespace/name instead of relying on the VMI runtime UUID.

This commit includes the needed changes in sanity/unit tests.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

---------

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>