Resize for extra capacity when restoring to greater target volume (#131)
* enable expanddisks gate

without this gate, the new storage is not made available to the VM
read more about this here:
x

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

* Resize for extra capacity when restoring to greater target volume

While we do allow restoring to a greater target volume,
this was a no-op for fs volumes, because we never resized the fs
to make use of the new capacity.

This change always resizes the fs if necessary on volume publish.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

* stop skipping test that is restoring to larger pvc

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

* Freeze linter version, install via go install

We're not ready for latest version atm

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

---------

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>