Ensure unpublish works for carry over VMIs (#134)
One upgrade consideration missed in PR 130 is that
ControllerUnpublish could be called against a VMI-level-hotplugged
carry over from an older driver version.
In that case, we would not issue a removevolume call
at all and fail in ensuring the volume got removed.

ControllerPublish doesn't suffer from a similar issue
since the volume would just get added on the VM level in addition.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>